# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  services_api-gateway_internal_handler_http.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      refresh_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      user:
        $ref: '#/definitions/services_api-gateway_internal_handler_http.UserInfo'
    type: object
...
  services_api-gateway_internal_handler_http.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
    type: object
...
  services_api-gateway_internal_handler_http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: mypassword123
        type: string
    type: object
...
  services_api-gateway_internal_handler_http.RefreshRequest:
    properties:
      refresh_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  services_api-gateway_internal_handler_http.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: Никита
        type: string
      last_name:
        example: Мозговой
        type: string
      password:
        example: mypassword123
        type: string
    type: object
...
  services_api-gateway_internal_handler_http.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      refresh_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  services_api-gateway_internal_handler_http.UserInfo:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: Никита
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_name:
        example: Мозговой
        type: string
      role:
        example: user
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the Delivery Tracker microservice platform
  title: Delivery Tracker API
  version: "1.0"
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, receive JWT tokens
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services_api-gateway_internal_handler_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.ErrorResponse'
      summary: Log in with credentials
      tags:
      - auth
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new token pair
      parameters:
      - description: Refresh payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services_api-gateway_internal_handler_http.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, first name and
        last name
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services_api-gateway_internal_handler_http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/services_api-gateway_internal_handler_http.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"